refactor(tomox): reorganize trading engine integration and update ChainContext references
